backend/services/auth-service/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
pyotp==2.9.0
Pillow==10.2.0
python-decouple==3.8
requests==2.31.0
# Email
django-ses==3.5.2
# SMS
twilio==8.12.0
# Testing
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
faker==23.2.0
coverage==7.4.1
